requirement.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
cryptocompare==0.7.6
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hurry==1.1
hurry.filesize==0.9
idna==3.4
Khayyam==3.0.17
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-decouple==3.8
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
TgCrypto==1.2.5
urllib3==2.0.4
